+++ src/main/java/hanghae/api/coupteambe/service/AuthKakaoService.java
@@ -1,0 +1,113 @@
+package hanghae.api.coupteambe.service;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import hanghae.api.coupteambe.domain.dto.social.SocialUserInfoDto;
+import hanghae.api.coupteambe.enumerate.Social;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+import org.springframework.web.client.RestTemplate;
+
+@Service
+@Slf4j
+public class AuthKakaoService {
+
+    public SocialUserInfoDto kakao(String code) throws JsonProcessingException {
+
+        //todo 프론트에서 받은 인가코드를 기반으로 인증서버에게 인증 받고,
+        // 인증받은 사용자의 정보를 이용하여 SocialUserInfoDto를 생성하여 반환한다.
+
+        //프론트에서 받은 인가코드를 기반으로 인증서버에게 인증 받고,
+        String accessToken = getAccessToken(code);
+
+        //인증받은 사용자의 정보를 이용하여 SocialUserInfoDto 를 생성하여 반환한다.
+        return getKakaoUserInfo(accessToken);
+    }
+
+    @Value("${auth.kakao.client-id}")
+    private String kakaoClientKId;
+
+    @Value("${auth.kakao.redirect-uri}")
+    private String kakaoRedirectUri;
+    // 1. "인가 코드"로 "액세스 토큰" 요청
+    private String getAccessToken(String code) throws JsonProcessingException {
+        // HTTP Header 생성
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
+
+        // HTTP Body 생성
+        MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
+        body.add("grant_type", "authorization_code");
+        body.add("client_id", kakaoClientKId);
+        body.add("redirect_uri", kakaoRedirectUri);
+        body.add("code", code);
+
+        // HTTP 요청 보내기
+        HttpEntity<MultiValueMap<String, String>> kakaoTokenRequest =
+                new HttpEntity<>(body, headers);
+        RestTemplate rt = new RestTemplate();
+        ResponseEntity<String> response = rt.exchange(
+                "https://kauth.kakao.com/oauth/token",
+                HttpMethod.POST,
+                kakaoTokenRequest,
+                String.class
+        );
+
+        // HTTP 응답 (JSON) -> 액세스 토큰 파싱
+        String responseBody = response.getBody();
+        ObjectMapper objectMapper = new ObjectMapper();
+        JsonNode jsonNode = objectMapper.readTree(responseBody);
+        String access_token = jsonNode.get("access_token").asText();
+        String refresh_token = jsonNode.get("refresh_token").asText();
+        log.debug("access token : " + access_token);
+        log.debug("refresh token : " + refresh_token);
+
+        return access_token;
+    }
+
+    // 2. "액세스 토큰"으로 "카카오 사용자 정보" 가져오기
+    private SocialUserInfoDto getKakaoUserInfo(String accessToken) throws JsonProcessingException {
+        // HTTP Header 생성
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Authorization", "Bearer " + accessToken);
+        headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
+
+        // HTTP 요청 보내기
+        HttpEntity<MultiValueMap<String, String>> kakaoUserInfoRequest = new HttpEntity<>(headers);
+        RestTemplate rt = new RestTemplate();
+        ResponseEntity<String> response = rt.exchange(
+                "https://kapi.kakao.com/v2/user/me",
+                HttpMethod.POST,
+                kakaoUserInfoRequest,
+                String.class
+        );
+
+        String responseBody = response.getBody();
+        ObjectMapper objectMapper = new ObjectMapper();
+        JsonNode jsonNode = objectMapper.readTree(responseBody);
+        Long id = jsonNode.get("id").asLong();
+        String nickname = jsonNode.get("properties").get("nickname").asText();
+        String email = jsonNode.get("kakao_account").get("email").asText();
+        String profile_image = jsonNode.get("properties").get("profile_image").asText();
+
+        log.debug("로그인 이용자 정보");
+        log.debug("카카오 고유ID : " + id);
+        log.debug("닉네임 : " + nickname);
+        log.debug("이메일 : " + email);
+        log.debug("프로필이미지 URL : " + profile_image);
+
+        return SocialUserInfoDto.builder()
+                         .loginId(email)
+                         .nickname(nickname)
+                         .profileImage(profile_image)
+                         .social(Social.KAKAO).build();
+    }
+}
